+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Change log for SPSWakeUp
 
-### 2.2.2
+### 2.1.3
+
+* Add functions:
+ - Get-SPWebServicesUrl
+ - Backup-HostsFile
+ - Restore-HostsFile
+ - Clear-HostsFileCopy
+* Add variables for HOSTS and Backup Hosts file with today's date
+ - hostsFile
+ - hostsFileCopy
+* Improved code styling by following PowerShell guidelines
+
+### 2.1.2
 
 * Add psVersion variable that contains the actual PowerShell version
 * Add function Get-SPSVersion
 * Check support between PSVersion and SharePoint Server 2010
 * Improved code styling by following PowerShell guidelines
 
-### 2.2.1
+### 2.1.1
 
 * Initial public release of SPSWakeUp on Github